AUDIT-64 Use device information from database (#442)
There are scenarios where the device information (specifically the vehicle_id) in the cache gets out of sync with the information in the database. This seems to happen on occasion when a provider updates a device's vehicle_id. There is a separate effort to diagnose that issue, but in the meantime, this fix will always use device information from the database which we know to be accurate.